Add settings to preset folder and permissions for journal entries
This adds a GM-only setting to determine what the default permissions
for journal entries created by the Map Pin creation will be, as well as
a setting to create said journal entries in a folder named after the
user creating the entry.

It also adds a helper when the perUser setting is saved, automatically
creating a folder per player. If a player does not have a folder, but
the option is set, a folder will be created when a Map Pin is created.