Foundry 9 Update Compatibility
Testing the latest version posted on Foundry 9 - Build 245, I noticed a few bugs.

Notably, trying to place more than one journal entry in a scene would result in the second marker not getting placed, and you not being able to place any more on the canvas until a page reload.

The cause of this seems to be how multiple times 'this.document.object.data' was used, and this was breaking things for reasons I don't understand. Changing these to 'this.document.getflag' seems to have fixed that issue entirely.

The await that was present at line 1204 was causing an error when placing a new marker stating that you must provide an _id for every new item in the update array. Commenting out that await cause the error to go away, and the marker will still update with the correct icon when hitting 'Update Map Note' to confirm placement anyway, so I don't see this as a huge loss.

The getFlag that was previously on line 1030 was returning as 'not a function', for whatever reason, but this is easily worked around by directly referencing the data in question.

Line 1039 has a simple change, from entities to documents. It stops a minor annoying warning in the console.

Finally, I checked compatibility with this for Foundry 0.8.9 and unfortunately it falls short. No horrible errors, but the 'show background?' checkbox in each marker does nothing at all, and the background is stuck as on. Because of that, I labeled this as a Foundry 9 exclusive update.